requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-openai
pinecone-client
supabase
torch
pydantic-settings
python-dotenv
tavily-python
openai